+++ neuron/src/lib/Neuron/Zettelkasten/Zettel/Parser.hs
@@ -12,7 +12,6 @@ import Control.Monad.Writer
 import Data.List (nub)
 import Data.Some
 import Data.TagTree (Tag)
-import qualified Data.Text as T
 import Neuron.Reader.Type
 import Neuron.Zettelkasten.ID
 import Neuron.Zettelkasten.Query.Error
@@ -39,9 +38,8 @@ parseZettel format zreader fn zid s = do
       let -- Determine zettel title
           (title, titleInBody) = case Meta.title =<< meta of
             Just tit -> (tit, False)
-            Nothing -> fromMaybe ("Untitled", False) $ do
+            Nothing -> fromMaybe (zettelIDRaw zid, False) $ do
               ((,True) . plainify . snd <$> getH1 doc)
-                <|> ((,False) . takeInitial . plainify <$> getFirstParagraphText doc)
           -- Accumulate queries
           (queries, errors) = runWriter $ extractQueries doc
           -- Determine zettel tags
@@ -68,8 +66,6 @@ parseZettel format zreader fn zid s = do
     getInlineTag = \case
       Some (ZettelQuery_TagZettel tag) -> Just tag
       _ -> Nothing
-    takeInitial =
-      (<> " ...") . T.take 18
 
 -- | Like `parseZettel` but operates on multiple files.
 parseZettels ::